Saga to update notifications permission status upon application entering foreground
