+++ ts/features/pushNotifications/sagas/__tests__/profileAndSystemNotificationsPermissions.test.tsx
@@ -2,12 +2,9 @@ import { CommonActions, StackActions } from "@react-navigation/native";
 import { testSaga } from "redux-saga-test-plan";
 import NavigationService from "../../../../navigation/NavigationService";
 import ROUTES from "../../../../navigation/routes";
-import {
-  checkNotificationPermissions,
-  requestNotificationPermissions
-} from "../../utils";
+import { requestNotificationPermissions } from "../../utils";
 import { notificationsInfoScreenConsent } from "../../store/actions/notifications";
-import { checkNotificationsPreferencesSaga } from "../checkNotificationsPreferencesSaga";
+import { profileAndSystemNotificationsPermissions } from "../profileAndSystemNotificationsPermissions";
 import { InitializedProfile } from "../../../../../definitions/backend/InitializedProfile";
 import { ServicesPreferencesModeEnum } from "../../../../../definitions/backend/ServicesPreferencesMode";
 import { profileUpsert } from "../../../../store/actions/profile";
@@ -19,7 +16,10 @@ import {
 } from "../../analytics";
 import { updateMixpanelSuperProperties } from "../../../../mixpanelConfig/superProperties";
 import { updateMixpanelProfileProperties } from "../../../../mixpanelConfig/profileProperties";
-import { updateSystemNotificationsEnabled } from "../../store/actions/permissions";
+import {
+  checkAndUpdateNotificationPermissionsIfNeeded,
+  updateNotificationPermissionsIfNeeded
+} from "../common";
 
 const generateUserProfile = (
   hasDoneNotificationOptIn: boolean,
@@ -53,7 +53,7 @@ describe("checkNotificationsPreferencesSaga", () => {
     const profile = generateUserProfile(false, true);
     const profileUpsertOutput = profileUpsertResult();
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -75,13 +75,11 @@ describe("checkNotificationsPreferencesSaga", () => {
         profileUpsertOutput.payload.newValue.reminder_status
       )
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
+      .call(updateNotificationPermissionsIfNeeded, true)
       .next()
       .select()
       .next(globalState)
@@ -97,7 +95,7 @@ describe("checkNotificationsPreferencesSaga", () => {
     const profile = generateUserProfile(false, true);
     const profileUpsertOutput = profileUpsertResult();
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -119,13 +117,11 @@ describe("checkNotificationsPreferencesSaga", () => {
         profileUpsertOutput.payload.newValue.reminder_status
       )
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
+      .call(updateNotificationPermissionsIfNeeded, false)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -152,7 +148,7 @@ describe("checkNotificationsPreferencesSaga", () => {
     const profile = generateUserProfile(false, true);
     const profileUpsertOutput = profileUpsertResult();
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -174,10 +170,8 @@ describe("checkNotificationsPreferencesSaga", () => {
         profileUpsertOutput.payload.newValue.reminder_status
       )
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
-      .next()
       .select()
       .next(globalState)
       .call(updateMixpanelSuperProperties, globalState)
@@ -191,15 +185,13 @@ describe("checkNotificationsPreferencesSaga", () => {
   it("profile has     notification settings, missing service configuration, device has no notification permissions, gives  device notification permissions", () => {
     const profile = generateUserProfile(true, true);
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
+      .call(updateNotificationPermissionsIfNeeded, true)
       .next()
       .select()
       .next(globalState)
@@ -212,15 +204,13 @@ describe("checkNotificationsPreferencesSaga", () => {
   it("profile has     notification settings, missing service configuration, device has no notification permissions, denies device notification permissions", () => {
     const profile = generateUserProfile(true, true);
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
+      .call(updateNotificationPermissionsIfNeeded, false)
       .next()
       .select()
       .next(globalState)
@@ -233,12 +223,10 @@ describe("checkNotificationsPreferencesSaga", () => {
   it("profile has     notification settings, missing service configuration, device has    notification permissions", () => {
     const profile = generateUserProfile(true, true);
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
-      .next()
       .select()
       .next(globalState)
       .call(updateMixpanelSuperProperties, globalState)
@@ -251,7 +239,7 @@ describe("checkNotificationsPreferencesSaga", () => {
     const profile = generateUserProfile(false, false);
     const profileUpsertOutput = profileUpsertResult();
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -273,13 +261,11 @@ describe("checkNotificationsPreferencesSaga", () => {
         profileUpsertOutput.payload.newValue.reminder_status
       )
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
+      .call(updateNotificationPermissionsIfNeeded, true)
       .next()
       .select()
       .next(globalState)
@@ -295,7 +281,7 @@ describe("checkNotificationsPreferencesSaga", () => {
     const profile = generateUserProfile(false, false);
     const profileUpsertOutput = profileUpsertResult();
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -317,13 +303,11 @@ describe("checkNotificationsPreferencesSaga", () => {
         profileUpsertOutput.payload.newValue.reminder_status
       )
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
+      .call(updateNotificationPermissionsIfNeeded, false)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -350,7 +334,7 @@ describe("checkNotificationsPreferencesSaga", () => {
     const profile = generateUserProfile(false, false);
     const profileUpsertOutput = profileUpsertResult();
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
       .call(
         NavigationService.dispatchNavigationAction,
@@ -372,10 +356,8 @@ describe("checkNotificationsPreferencesSaga", () => {
         profileUpsertOutput.payload.newValue.reminder_status
       )
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
-      .next()
       .select()
       .next(globalState)
       .call(updateMixpanelSuperProperties, globalState)
@@ -389,15 +371,13 @@ describe("checkNotificationsPreferencesSaga", () => {
   it("profile has     notification settings, has     service configuration, device has no notification permissions, gives  device notification permissions", () => {
     const profile = generateUserProfile(true, false);
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
+      .call(updateNotificationPermissionsIfNeeded, true)
       .next()
       .select()
       .next(globalState)
@@ -410,15 +390,13 @@ describe("checkNotificationsPreferencesSaga", () => {
   it("profile has     notification settings, has     service configuration, device has no notification permissions, denies device notification permissions", () => {
     const profile = generateUserProfile(true, false);
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
-      .next()
       .call(requestNotificationPermissions)
       .next(false)
-      .put(updateSystemNotificationsEnabled(false))
+      .call(updateNotificationPermissionsIfNeeded, false)
       .next()
       .select()
       .next(globalState)
@@ -431,12 +409,10 @@ describe("checkNotificationsPreferencesSaga", () => {
   it("profile has     notification settings, has     service configuration, device has    notification permissions", () => {
     const profile = generateUserProfile(true, false);
     const globalState = {};
-    testSaga(checkNotificationsPreferencesSaga, profile)
+    testSaga(profileAndSystemNotificationsPermissions, profile)
       .next()
-      .call(checkNotificationPermissions)
+      .call(checkAndUpdateNotificationPermissionsIfNeeded)
       .next(true)
-      .put(updateSystemNotificationsEnabled(true))
-      .next()
       .select()
       .next(globalState)
       .call(updateMixpanelSuperProperties, globalState)
