+++ ts/sagas/__tests__/initializeApplicationSaga.test.ts
@@ -42,6 +42,7 @@ import { watchLogoutSaga } from "../startup/watchLogoutSaga";
 import { cancellAllLocalNotifications } from "../../features/pushNotifications/utils";
 import { handleApplicationStartupTransientError } from "../../features/startup/sagas";
 import { startupTransientErrorInitialState } from "../../store/reducers/startup";
+import { notificationPermissionsListener } from "../../features/pushNotifications/sagas/notificationPermissionsListener";
 
 const aSessionToken = "a_session_token" as SessionToken;
 
@@ -80,6 +81,8 @@ describe("initializeApplicationSaga", () => {
       .put(previousInstallationDataDeleteSuccess())
       .next()
       .next()
+      .next()
+      .fork(notificationPermissionsListener)
       .next()
       .select(profileSelector)
       .next(pot.some(profile))
@@ -130,6 +133,8 @@ describe("initializeApplicationSaga", () => {
       .next()
       .next()
       .next()
+      .fork(notificationPermissionsListener)
+      .next()
       .select(profileSelector)
       .next(pot.some(profile))
       .fork(watchProfileEmailValidationChangedSaga, O.none)
@@ -171,6 +176,8 @@ describe("initializeApplicationSaga", () => {
       .put(previousInstallationDataDeleteSuccess())
       .next()
       .next()
+      .next()
+      .fork(notificationPermissionsListener)
       .next()
       .select(profileSelector)
       .next(pot.some(profile))
@@ -218,6 +225,8 @@ describe("initializeApplicationSaga", () => {
       .put(previousInstallationDataDeleteSuccess())
       .next()
       .next()
+      .next()
+      .fork(notificationPermissionsListener)
       .next()
       .select(profileSelector)
       .next(pot.some(profile))
@@ -278,6 +287,8 @@ describe("initializeApplicationSaga", () => {
       .next()
       .next()
       .next()
+      .fork(notificationPermissionsListener)
+      .next()
       .select(profileSelector)
       .next(pot.some(profile))
       .fork(watchProfileEmailValidationChangedSaga, O.none)
@@ -325,6 +336,8 @@ describe("initializeApplicationSaga", () => {
       .put(previousInstallationDataDeleteSuccess())
       .next()
       .next()
+      .next()
+      .fork(notificationPermissionsListener)
       .next()
       .select(profileSelector)
       .next(pot.some(profile))
