+++ ts/sagas/startup.ts
@@ -108,11 +108,10 @@ import { watchWalletSaga as watchNewWalletSaga } from "../features/newWallet/sag
 import { watchServicesSaga } from "../features/services/common/saga";
 import { watchItwSaga } from "../features/itwallet/common/saga";
 import { watchTrialSystemSaga } from "../features/trialSystem/store/sagas/watchTrialSystemSaga";
-import {
-  handlePendingMessageStateIfAllowedSaga,
-  updateInstallationSaga
-} from "../features/pushNotifications/sagas/notifications";
-import { checkNotificationsPreferencesSaga } from "../features/pushNotifications/sagas/checkNotificationsPreferencesSaga";
+import { notificationPermissionsListener } from "../features/pushNotifications/sagas/notificationPermissionsListener";
+import { profileAndSystemNotificationsPermissions } from "../features/pushNotifications/sagas/profileAndSystemNotificationsPermissions";
+import { pushNotificationTokenUpload } from "../features/pushNotifications/sagas/pushNotificationTokenUpload";
+import { handlePendingMessageStateIfAllowed } from "../features/pushNotifications/sagas/common";
 import { cancellAllLocalNotifications } from "../features/pushNotifications/utils";
 import { handleApplicationStartupTransientError } from "../features/startup/sagas";
 import {
@@ -203,6 +202,9 @@ export function* initializeApplicationSaga(
 
   // clear cached downloads when the logged user changes
   yield* takeEvery(differentProfileLoggedIn, handleClearAllAttachments);
+
+  // Retrieve and listen for notification permissions status changes
+  yield* fork(notificationPermissionsListener);
 
   // Get last logged in Profile from the state
   const lastLoggedInProfileState: ReturnType<typeof profileSelector> =
@@ -509,7 +511,7 @@ export function* initializeApplicationSaga(
   userProfile = (yield* call(checkEmailSaga)) ?? userProfile;
 
   // check if the user must set preferences for push notifications (e.g. reminders)
-  yield* call(checkNotificationsPreferencesSaga, userProfile);
+  yield* call(profileAndSystemNotificationsPermissions, userProfile);
 
   const isFirstOnboarding = isProfileFirstOnBoarding(userProfile);
   yield* call(askServicesPreferencesModeOptin, isFirstOnboarding);
@@ -524,7 +526,10 @@ export function* initializeApplicationSaga(
 
   // Start the notification installation update as early as
   // possible to begin receiving push notifications
-  yield* call(updateInstallationSaga, backendClient.createOrUpdateInstallation);
+  yield* call(
+    pushNotificationTokenUpload,
+    backendClient.createOrUpdateInstallation
+  );
 
   // This saga is called before the startup status is set to authenticated to avoid flashing
   // the home screen when the user is taken to the alert screen in case of identities that don't match.
@@ -666,7 +671,7 @@ export function* initializeApplicationSaga(
   yield* fork(watchEmailNotificationPreferencesSaga);
 
   // Check if we have a pending notification message
-  yield* call(handlePendingMessageStateIfAllowedSaga, true);
+  yield* call(handlePendingMessageStateIfAllowed, true);
 
   // This tells the security advice bottomsheet that it can be shown
   yield* put(setSecurityAdviceReadyToShow(true));
